# openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
...
paths:
...
  /office/:
    put:
      tags:
        - Офис
      summary: Изменить указанный офис
      description: Все поля обязательны
      operationId: update_4
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OfficeUpdateDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: string
        '400':
          description: >-
            Ошибки валидации. Два варианта возвращаемого JSON объекта (смотреть
            Examples)
          content:
            '*/*':
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ExceptionDto'
                  - $ref: '#/components/schemas/ExceptionsDto'
              examples:
                Вариант с одной ошибкой:
                  description: Вариант с одной ошибкой
                  value:
                    status: BAD_REQUEST
                    timestamp: 08-08-2008 08:08:08
                    message: Что-либо не прошло валидацию.
                Вариант со списком ошибок:
                  description: Вариант со списком ошибок
                  value:
                    status: BAD_REQUEST
                    timestamp: 08-08-2008 08:08:08
                    errors:
                      field: Поле field не прошло валидацию.
                      name: Поле name не прошло валидацию.
        '404':
          description: Объект не найден
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ExceptionDto'
        '409':
          description: Объект уже существует
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ExceptionDto'
    post:
      tags:
        - Офис
      summary: Создать офис
      description: Все поля обязательны
      operationId: createOffice
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OfficeCreateDto'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                type: integer
                format: int64
        '400':
          description: >-
            Ошибки валидации. Два варианта возвращаемого JSON объекта (смотреть
            Examples)
          content:
            '*/*':
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ExceptionDto'
                  - $ref: '#/components/schemas/ExceptionsDto'
              examples:
                Вариант с одной ошибкой:
                  description: Вариант с одной ошибкой
                  value:
                    status: BAD_REQUEST
                    timestamp: 08-08-2008 08:08:08
                    message: Что-либо не прошло валидацию.
                Вариант со списком ошибок:
                  description: Вариант со списком ошибок
                  value:
                    status: BAD_REQUEST
                    timestamp: 08-08-2008 08:08:08
                    errors:
                      field: Поле field не прошло валидацию.
                      name: Поле name не прошло валидацию.
        '404':
          description: Объект не найден
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ExceptionDto'
        '409':
          description: Объект уже существует
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ExceptionDto'
...
  /office/{id}:
    get:
      tags:
        - Офис
      summary: Получить указанный офис
      operationId: getOfficeById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/OfficeGetDto'
        '400':
          description: >-
            Ошибки валидации. Два варианта возвращаемого JSON объекта (смотреть
            Examples)
          content:
            '*/*':
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ExceptionDto'
                  - $ref: '#/components/schemas/ExceptionsDto'
              examples:
                Вариант с одной ошибкой:
                  description: Вариант с одной ошибкой
                  value:
                    status: BAD_REQUEST
                    timestamp: 08-08-2008 08:08:08
                    message: Что-либо не прошло валидацию.
                Вариант со списком ошибок:
                  description: Вариант со списком ошибок
                  value:
                    status: BAD_REQUEST
                    timestamp: 08-08-2008 08:08:08
                    errors:
                      field: Поле field не прошло валидацию.
                      name: Поле name не прошло валидацию.
        '404':
          description: Объект не найден
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ExceptionDto'
        '409':
          description: Объект уже существует
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ExceptionDto'
    delete:
      tags:
        - Офис
      summary: Удалить указанный офис
      operationId: delete_5
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: string
        '400':
          description: >-
            Ошибки валидации. Два варианта возвращаемого JSON объекта (смотреть
            Examples)
          content:
            '*/*':
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ExceptionDto'
                  - $ref: '#/components/schemas/ExceptionsDto'
              examples:
                Вариант с одной ошибкой:
                  description: Вариант с одной ошибкой
                  value:
                    status: BAD_REQUEST
                    timestamp: 08-08-2008 08:08:08
                    message: Что-либо не прошло валидацию.
                Вариант со списком ошибок:
                  description: Вариант со списком ошибок
                  value:
                    status: BAD_REQUEST
                    timestamp: 08-08-2008 08:08:08
                    errors:
                      field: Поле field не прошло валидацию.
                      name: Поле name не прошло валидацию.
        '404':
          description: Объект не найден
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ExceptionDto'
        '409':
          description: Объект уже существует
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ExceptionDto'
  /office/all:
    get:
      tags:
        - Офис
      summary: Все офисы
      description: 1 <= size <= 20 (default 20)
      operationId: getOffices
      parameters:
        - name: page
          in: query
          description: Zero-based page index (0..N)
          required: false
          schema:
            minimum: 0
            type: integer
            default: 0
        - name: size
          in: query
          description: The size of the page to be returned
          required: false
          schema:
            minimum: 1
            type: integer
            default: 20
        - name: sort
          in: query
          description: >-
            Sorting criteria in the format: property,(asc|desc). Default sort
            order is ascending. Multiple sort criteria are supported.
          required: false
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageOfficeGetDto'
        '400':
          description: >-
            Ошибки валидации. Два варианта возвращаемого JSON объекта (смотреть
            Examples)
          content:
            '*/*':
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ExceptionDto'
                  - $ref: '#/components/schemas/ExceptionsDto'
              examples:
                Вариант с одной ошибкой:
                  description: Вариант с одной ошибкой
                  value:
                    status: BAD_REQUEST
                    timestamp: 08-08-2008 08:08:08
                    message: Что-либо не прошло валидацию.
                Вариант со списком ошибок:
                  description: Вариант со списком ошибок
                  value:
                    status: BAD_REQUEST
                    timestamp: 08-08-2008 08:08:08
                    errors:
                      field: Поле field не прошло валидацию.
                      name: Поле name не прошло валидацию.
        '404':
          description: Объект не найден
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ExceptionDto'
        '409':
          description: Объект уже существует
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ExceptionDto'
  /office/allByAddress:
    get:
      tags:
        - Офис
      summary: Поиск среди офисов по адресу
      description: 1 <= size <= 20 (default 20)
      operationId: getOfficesByAddress
      parameters:
        - name: address
          in: query
          required: true
          schema:
            type: string
        - name: page
          in: query
          description: Zero-based page index (0..N)
          required: false
          schema:
            minimum: 0
            type: integer
            default: 0
        - name: size
          in: query
          description: The size of the page to be returned
          required: false
          schema:
            minimum: 1
            type: integer
            default: 20
        - name: sort
          in: query
          description: >-
            Sorting criteria in the format: property,(asc|desc). Default sort
            order is ascending. Multiple sort criteria are supported.
          required: false
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageOfficeGetDto'
        '400':
          description: >-
            Ошибки валидации. Два варианта возвращаемого JSON объекта (смотреть
            Examples)
          content:
            '*/*':
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ExceptionDto'
                  - $ref: '#/components/schemas/ExceptionsDto'
              examples:
                Вариант с одной ошибкой:
                  description: Вариант с одной ошибкой
                  value:
                    status: BAD_REQUEST
                    timestamp: 08-08-2008 08:08:08
                    message: Что-либо не прошло валидацию.
                Вариант со списком ошибок:
                  description: Вариант со списком ошибок
                  value:
                    status: BAD_REQUEST
                    timestamp: 08-08-2008 08:08:08
                    errors:
                      field: Поле field не прошло валидацию.
                      name: Поле name не прошло валидацию.
        '404':
          description: Объект не найден
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ExceptionDto'
        '409':
          description: Объект уже существует
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ExceptionDto'
...
  /booking/office/:
    get:
      tags:
        - Бронирования
      summary: Получение страницы броней по офису (включая удаленные)
      operationId: getAllOfOffice
      parameters:
        - name: officeId
          in: query
          required: true
          schema:
            type: integer
            format: int64
        - name: typeId
          in: query
          required: true
          schema:
            type: integer
            format: int64
        - name: isActual
          in: query
          description: >-
            Параметр отвечает за выборку броней: актуальные - все. Параметр не
            обязателен, дефолтное значение - true.
          required: false
          schema:
            type: boolean
            default: true
        - name: page
          in: query
          description: Zero-based page index (0..N)
          required: false
          schema:
            minimum: 0
            type: integer
            default: 0
        - name: size
          in: query
          description: The size of the page to be returned
          required: false
          schema:
            minimum: 1
            type: integer
            default: 20
        - name: sort
          in: query
          description: >-
            Sorting criteria in the format: property,(asc|desc). Default sort
            order is ascending. Multiple sort criteria are supported.
          required: false
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageBookingGetDto'
        '400':
          description: >-
            Ошибки валидации. Два варианта возвращаемого JSON объекта (смотреть
            Examples)
          content:
            '*/*':
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ExceptionDto'
                  - $ref: '#/components/schemas/ExceptionsDto'
              examples:
                Вариант с одной ошибкой:
                  description: Вариант с одной ошибкой
                  value:
                    status: BAD_REQUEST
                    timestamp: 08-08-2008 08:08:08
                    message: Что-либо не прошло валидацию.
                Вариант со списком ошибок:
                  description: Вариант со списком ошибок
                  value:
                    status: BAD_REQUEST
                    timestamp: 08-08-2008 08:08:08
                    errors:
                      field: Поле field не прошло валидацию.
                      name: Поле name не прошло валидацию.
        '404':
          description: Объект не найден
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ExceptionDto'
        '409':
          description: Объект уже существует
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ExceptionDto'
...
components:
  schemas:
    ExceptionDto:
      type: object
      properties:
        status:
          type: string
          enum:
            - 100 CONTINUE
            - 101 SWITCHING_PROTOCOLS
            - 102 PROCESSING
            - 103 CHECKPOINT
            - 200 OK
            - 201 CREATED
            - 202 ACCEPTED
            - 203 NON_AUTHORITATIVE_INFORMATION
            - 204 NO_CONTENT
            - 205 RESET_CONTENT
            - 206 PARTIAL_CONTENT
            - 207 MULTI_STATUS
            - 208 ALREADY_REPORTED
            - 226 IM_USED
            - 300 MULTIPLE_CHOICES
            - 301 MOVED_PERMANENTLY
            - 302 FOUND
            - 302 MOVED_TEMPORARILY
            - 303 SEE_OTHER
            - 304 NOT_MODIFIED
            - 305 USE_PROXY
            - 307 TEMPORARY_REDIRECT
            - 308 PERMANENT_REDIRECT
            - 400 BAD_REQUEST
            - 401 UNAUTHORIZED
            - 402 PAYMENT_REQUIRED
            - 403 FORBIDDEN
            - 404 NOT_FOUND
            - 405 METHOD_NOT_ALLOWED
            - 406 NOT_ACCEPTABLE
            - 407 PROXY_AUTHENTICATION_REQUIRED
            - 408 REQUEST_TIMEOUT
            - 409 CONFLICT
            - 410 GONE
            - 411 LENGTH_REQUIRED
            - 412 PRECONDITION_FAILED
            - 413 PAYLOAD_TOO_LARGE
            - 413 REQUEST_ENTITY_TOO_LARGE
            - 414 URI_TOO_LONG
            - 414 REQUEST_URI_TOO_LONG
            - 415 UNSUPPORTED_MEDIA_TYPE
            - 416 REQUESTED_RANGE_NOT_SATISFIABLE
            - 417 EXPECTATION_FAILED
            - 418 I_AM_A_TEAPOT
            - 419 INSUFFICIENT_SPACE_ON_RESOURCE
            - 420 METHOD_FAILURE
            - 421 DESTINATION_LOCKED
            - 422 UNPROCESSABLE_ENTITY
            - 423 LOCKED
            - 424 FAILED_DEPENDENCY
            - 425 TOO_EARLY
            - 426 UPGRADE_REQUIRED
            - 428 PRECONDITION_REQUIRED
            - 429 TOO_MANY_REQUESTS
            - 431 REQUEST_HEADER_FIELDS_TOO_LARGE
            - 451 UNAVAILABLE_FOR_LEGAL_REASONS
            - 500 INTERNAL_SERVER_ERROR
            - 501 NOT_IMPLEMENTED
            - 502 BAD_GATEWAY
            - 503 SERVICE_UNAVAILABLE
            - 504 GATEWAY_TIMEOUT
            - 505 HTTP_VERSION_NOT_SUPPORTED
            - 506 VARIANT_ALSO_NEGOTIATES
            - 507 INSUFFICIENT_STORAGE
            - 508 LOOP_DETECTED
            - 509 BANDWIDTH_LIMIT_EXCEEDED
            - 510 NOT_EXTENDED
            - 511 NETWORK_AUTHENTICATION_REQUIRED
        timestamp:
          type: string
          format: date-time
        message:
          type: string
    ExceptionsDto:
      type: object
      properties:
        status:
          type: string
          enum:
            - 100 CONTINUE
            - 101 SWITCHING_PROTOCOLS
            - 102 PROCESSING
            - 103 CHECKPOINT
            - 200 OK
            - 201 CREATED
            - 202 ACCEPTED
            - 203 NON_AUTHORITATIVE_INFORMATION
            - 204 NO_CONTENT
            - 205 RESET_CONTENT
            - 206 PARTIAL_CONTENT
            - 207 MULTI_STATUS
            - 208 ALREADY_REPORTED
            - 226 IM_USED
            - 300 MULTIPLE_CHOICES
            - 301 MOVED_PERMANENTLY
            - 302 FOUND
            - 302 MOVED_TEMPORARILY
            - 303 SEE_OTHER
            - 304 NOT_MODIFIED
            - 305 USE_PROXY
            - 307 TEMPORARY_REDIRECT
            - 308 PERMANENT_REDIRECT
            - 400 BAD_REQUEST
            - 401 UNAUTHORIZED
            - 402 PAYMENT_REQUIRED
            - 403 FORBIDDEN
            - 404 NOT_FOUND
            - 405 METHOD_NOT_ALLOWED
            - 406 NOT_ACCEPTABLE
            - 407 PROXY_AUTHENTICATION_REQUIRED
            - 408 REQUEST_TIMEOUT
            - 409 CONFLICT
            - 410 GONE
            - 411 LENGTH_REQUIRED
            - 412 PRECONDITION_FAILED
            - 413 PAYLOAD_TOO_LARGE
            - 413 REQUEST_ENTITY_TOO_LARGE
            - 414 URI_TOO_LONG
            - 414 REQUEST_URI_TOO_LONG
            - 415 UNSUPPORTED_MEDIA_TYPE
            - 416 REQUESTED_RANGE_NOT_SATISFIABLE
            - 417 EXPECTATION_FAILED
            - 418 I_AM_A_TEAPOT
            - 419 INSUFFICIENT_SPACE_ON_RESOURCE
            - 420 METHOD_FAILURE
            - 421 DESTINATION_LOCKED
            - 422 UNPROCESSABLE_ENTITY
            - 423 LOCKED
            - 424 FAILED_DEPENDENCY
            - 425 TOO_EARLY
            - 426 UPGRADE_REQUIRED
            - 428 PRECONDITION_REQUIRED
            - 429 TOO_MANY_REQUESTS
            - 431 REQUEST_HEADER_FIELDS_TOO_LARGE
            - 451 UNAVAILABLE_FOR_LEGAL_REASONS
            - 500 INTERNAL_SERVER_ERROR
            - 501 NOT_IMPLEMENTED
            - 502 BAD_GATEWAY
            - 503 SERVICE_UNAVAILABLE
            - 504 GATEWAY_TIMEOUT
            - 505 HTTP_VERSION_NOT_SUPPORTED
            - 506 VARIANT_ALSO_NEGOTIATES
            - 507 INSUFFICIENT_STORAGE
            - 508 LOOP_DETECTED
            - 509 BANDWIDTH_LIMIT_EXCEEDED
            - 510 NOT_EXTENDED
            - 511 NETWORK_AUTHENTICATION_REQUIRED
        timestamp:
          type: string
          format: date-time
        errors:
          type: object
          additionalProperties:
            type: string
...
    OfficeUpdateDto:
      required:
        - address
        - bookingRange
        - cityId
        - endOfDay
        - id
        - startOfDay
        - workNumber
      type: object
      properties:
        id:
          minimum: 1
          type: integer
          format: int64
        cityId:
          minimum: 1
          type: integer
          format: int64
        address:
          type: string
        workNumber:
          pattern: ^((8|\+7)[\- ]?)?(\(?\d{3}\)?[\- ]?)?[\d\- ]{7,10}$
          type: string
        startOfDay:
          type: string
          format: hh:mm
          example: '00:00'
        endOfDay:
          type: string
          format: hh:mm
          example: '00:00'
        bookingRange:
          minimum: 1
          type: integer
          format: int32
...
    OfficeCreateDto:
      required:
        - address
        - bookingRange
        - cityId
        - endOfDay
        - startOfDay
        - workNumber
      type: object
      properties:
        cityId:
          minimum: 1
          type: integer
          format: int64
        address:
          type: string
        workNumber:
          pattern: ^((8|\+7)[\- ]?)?(\(?\d{3}\)?[\- ]?)?[\d\- ]{7,10}$
          type: string
        startOfDay:
          type: string
          format: hh:mm
          example: '00:00'
        endOfDay:
          type: string
          format: hh:mm
          example: '00:00'
        bookingRange:
          minimum: 1
          type: integer
          format: int32
...
    LocalTime:
      type: object
      properties:
        hour:
          type: integer
          format: int32
        minute:
          type: integer
          format: int32
        second:
          type: integer
          format: int32
        nano:
          type: integer
          format: int32
...
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          $ref: '#/components/schemas/SortObject'
        pageSize:
          type: integer
          format: int32
        pageNumber:
          type: integer
          format: int32
        unpaged:
          type: boolean
        paged:
          type: boolean
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean
...
    BookingGetDto:
      type: object
      properties:
        infoBooking:
          $ref: '#/components/schemas/InfoBookingDto'
        infoPlace:
          $ref: '#/components/schemas/InfoPlaceDto'
        infoOffice:
          $ref: '#/components/schemas/InfoOfficeDto'
...
    InfoBookingDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        makerId:
          type: integer
          format: int64
        makerName:
          type: string
        holderId:
          type: integer
          format: int64
        holderName:
          type: string
        start:
          type: string
          format: date-time
        end:
          type: string
          format: date-time
        guests:
          type: integer
          format: int32
        isDeleted:
          type: boolean
    InfoOfficeDto:
      type: object
      properties:
        officeId:
          type: integer
          format: int64
        city:
          type: string
        zoneId:
          type: string
        address:
          type: string
    InfoPlaceDto:
      type: object
      properties:
        placeId:
          type: integer
          format: int64
        placeName:
          type: string
        type:
          type: string
        floorNumber:
          type: integer
          format: int32
...
    OfficeGetDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        cityId:
          type: integer
          format: int64
        cityName:
          type: string
        zoneId:
          type: string
        address:
          type: string
        workNumber:
          type: string
        startOfDay:
          $ref: '#/components/schemas/LocalTime'
        endOfDay:
          $ref: '#/components/schemas/LocalTime'
        bookingRange:
          type: integer
          format: int32
    PageOfficeGetDto:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/OfficeGetDto'
        number:
          type: integer
          format: int32
        sort:
          $ref: '#/components/schemas/SortObject'
        numberOfElements:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        pageable:
          $ref: '#/components/schemas/PageableObject'
        empty:
          type: boolean
...
    PageBookingGetDto:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/BookingGetDto'
        number:
          type: integer
          format: int32
        sort:
          $ref: '#/components/schemas/SortObject'
        numberOfElements:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        pageable:
          $ref: '#/components/schemas/PageableObject'
        empty:
          type: boolean
...